[Refactor] 폴더 구조 리펙토링
