+++ src/pages/calendar/index.ts
@@ -1,0 +1,1 @@
+export { CalendarPage } from './ui/page'
